refactor(io): extract shared Arrow writer base for Lance and Vortex
